refactored code + style fixes
